feat: validações de ids enviados

+++ src/endpoints/accounts-receivable/accounts-receivable.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateAccountsReceivableDto } from './dto/create-accounts-receivable.dto';
 import { UpdateAccountsReceivableDto } from './dto/update-accounts-receivable.dto';
 import { PrismaClient } from '@prisma/client';
@@ -6,10 +6,24 @@ import { PrismaClient } from '@prisma/client';
 const prisma = new PrismaClient();
 @Injectable()
 export class AccountsReceivableService {
-  create(createAccountsReceivableDto: CreateAccountsReceivableDto) {
+  async create(createAccountsReceivableDto: CreateAccountsReceivableDto) {
+    const sheet = await prisma.sheets.findUnique({ where: { id: createAccountsReceivableDto.sheetId } });
+
+    if (!sheet) {
+      throw new HttpException(`not found sheetId: ${createAccountsReceivableDto.sheetId}`, HttpStatus.NOT_FOUND);
+    }
+
+    const user = await prisma.users.findUnique({ where: { id: createAccountsReceivableDto.creatorUserId } });
+    
+    if (!user) {
+      throw new HttpException(`not found userId: ${createAccountsReceivableDto.creatorUserId}`, HttpStatus.NOT_FOUND);
+    }
+
     return prisma.accountsReceivable.create({
       data: {
         ...createAccountsReceivableDto,
+        creatorUserId: user.id,
+        sheetId: sheet.id,
         createdAt: new Date(),
         updatedAt: new Date()
       }
@@ -41,8 +55,8 @@ export class AccountsReceivableService {
     });
   }
 
-  findOne(id: number) {
-    return prisma.accountsReceivable.findUnique({
+  async findOne(id: number) {
+    const accountReceivable = await prisma.accountsReceivable.findUnique({
       where: {
         id
       },
@@ -67,9 +81,25 @@ export class AccountsReceivableService {
         }
       }
     });;
+
+    if (!accountReceivable) {
+      throw new HttpException(`not found accountReceivableId: ${id}`, HttpStatus.NOT_FOUND);
+    }
+
+    return accountReceivable;
   }
 
-  update(id: number, updateAccountsReceivableDto: UpdateAccountsReceivableDto) {
+  async update(id: number, updateAccountsReceivableDto: UpdateAccountsReceivableDto) {
+    const accountReceivable = await prisma.accountsReceivable.findUnique({
+      where: {
+        id
+      },
+    });;
+
+    if (!accountReceivable) {
+      throw new HttpException(`not found accountReceivableId: ${id}`, HttpStatus.NOT_FOUND);
+    }
+
     return prisma.accountsReceivable.update({
       where: {
         id
@@ -81,7 +111,17 @@ export class AccountsReceivableService {
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const accountReceivable = await prisma.accountsReceivable.findUnique({
+      where: {
+        id
+      },
+    });;
+
+    if (!accountReceivable) {
+      throw new HttpException(`not found accountReceivableId: ${id}`, HttpStatus.NOT_FOUND);
+    }
+
     return prisma.accountsReceivable.delete({
       where: {
         id
